requirements.txt
pymongo>=4.6
PyYAML>=6.0
matplotlib>=3.7


Flask>=3.0
